Remove unneeded outputs from shell_exec

+++ src/Phug/Split/Command/CommandBase.php
@@ -109,6 +109,10 @@ abstract class CommandBase extends Command
      */
     protected function remove(string $fileOrDirectory): bool
     {
+        if (!is_writable($fileOrDirectory)) {
+            return false;
+        }
+
         @shell_exec('rm -rf '.escapeshellarg($fileOrDirectory).' 2>&1');
         file_exists($fileOrDirectory) && @shell_exec('rmdir /S /Q '.escapeshellarg($fileOrDirectory).' 2>&1');
